fixed64: codemod pass 7 -- mechanical invalid-operands (double/compound)
double<->fixed operands cast the Fixed side to double(); compound-assign
of a Fixed into an int/angle accumulator (e.g. forward += forwardmove[speed])
casts the Fixed to the target type. 30 sites across 10 files, behavior-
preserving. The Fixed*Fixed / Fixed/Fixed multiplies are intentionally
skipped for per-site hand-fixing next. Flag-off green; tests green.

+++ src/zandronum/src/g_game.cpp
@@ -756,9 +756,9 @@ void G_BuildTiccmd (ticcmd_t *cmd)
 	if (strafe)
 	{
 		if (Button_Right.bDown)
-			side += sidemove[speed];
+			side += (int)(sidemove[speed]);
 		if (Button_Left.bDown)
-			side -= sidemove[speed];
+			side -= (int)(sidemove[speed]);
 	}
 	else
 	{
@@ -787,9 +787,9 @@ void G_BuildTiccmd (ticcmd_t *cmd)
 	}
 
 	if (Button_MoveUp.bDown)
-		fly += flyspeed[speed];
+		fly += (int)(flyspeed[speed]);
 	if (Button_MoveDown.bDown)
-		fly -= flyspeed[speed];
+		fly -= (int)(flyspeed[speed]);
 
 	if (Button_Klook.bDown)
 	{
@@ -801,15 +801,15 @@ void G_BuildTiccmd (ticcmd_t *cmd)
 	else
 	{
 		if (Button_Forward.bDown)
-			forward += forwardmove[speed];
+			forward += (int)(forwardmove[speed]);
 		if (Button_Back.bDown)
-			forward -= forwardmove[speed];
+			forward -= (int)(forwardmove[speed]);
 	}
 
 	if (Button_MoveRight.bDown)
-		side += sidemove[speed];
+		side += (int)(sidemove[speed]);
 	if (Button_MoveLeft.bDown)
-		side -= sidemove[speed];
+		side -= (int)(sidemove[speed]);
 
 	// buttons
 	if (Button_Attack.bDown)		cmd->ucmd.buttons |= BT_ATTACK;
